+++ packages/sdk-metrics/test/performance/benchmark/otel-sdk-metrics-paths.js
